reduece number of files
