--- errorcode.h
@@ -1,7 +1,0 @@
-#ifndef _ERRORCODE_H_
-#define _ERRORCODE_H_
-
-#define NULL_PTR -1
-#define WRONG_SPECIFIER -2
-
-#endif
